+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Rejected-PEPs Changelog
 
+### <u>0.6.1</u>  _Final_
+
+_Release Date: 2021-08-22_
+
+#### Fixed
+
+- Fixed a possible `UnboundLocalError` in `pep294.apply()`.
+
+#### Improved
+
+- `to_roman(x)` now checks for `x.__index__` before falling back to `int(x)` & `int(x, 0)`.
+
 ### <u>0.6.0</u>  _Final_
 
 _Release Date: 2021-08-22_